+++ changelog.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `feature` Add link-items for post-events publications.
 - `feature` Add memories in frontpage.
 - `feature` Add service for supported events (TODO).
+- `feature` Add SourceName for link-items.
 
 ### Fix